+++ modules/TargetHealth.lua
@@ -1012,12 +1012,13 @@ end
 function IceTargetHealth.prototype:CheckPvP()
 	local pvpMode = nil
 	local minx, maxx, miny, maxy
+	local unitIsPVP = UnitIsPVP(self.unit)
 
 	if self.configMode or UnitIsPVPFreeForAll(self.unit) then
 		pvpMode = "FFA"
 
 		minx, maxx, miny, maxy = 0.05, 0.605, 0.015, 0.57
-	elseif UnitIsPVP(self.unit) then
+	elseif IceHUD.CanAccessValue(unitIsPVP) and unitIsPVP then
 		pvpMode = UnitFactionGroup(self.unit)
 
 		if pvpMode == "Neutral" then